fixed messages and file plugins for #3544

+++ mod/messages/pages/messages/send.php
@@ -14,7 +14,7 @@ $title = elgg_echo('messages:add');
 
 elgg_push_breadcrumb($title);
 
-$params = messages_prepare_form_vars(get_input('send_to'));
+$params = messages_prepare_form_vars((int)get_input('send_to'));
 $params['friends'] = $page_owner->getFriends();
 $content = elgg_view_form('messages/send', array(), $params);
